refactor: split compile pipeline stages
