+++ src/pipeline.rs
@@ -10,12 +10,12 @@
 //! - Check/EIR/assembly-only paths return before read-only native artifact resolution.
 
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process;
 use std::time::Instant;
 
 use crate::cli::CliConfig;
-use crate::codegen::platform::{Platform, Target};
+use crate::codegen::platform::Target;
 use crate::codegen::Emit;
 use crate::codegen::LinkRequirement;
 use crate::native_deps::NativeRequirement;
@@ -28,13 +28,12 @@ use crate::{
     resolver, runtime_cache, source_map, tz_prelude, types, var_export_prelude, web_prelude,
 };
 
-/// Holds the paths for all compilation output files (assembly, object, binary, source map).
-struct OutputPaths {
-    asm: PathBuf,
-    obj: PathBuf,
-    bin: PathBuf,
-    source_map: PathBuf,
-}
+mod backend;
+mod eir_output;
+mod frontend;
+mod output;
+
+use output::{dynamic_eval_capability_warning, output_paths, OutputPaths};
 
 /// Runs the full compilation pipeline from PHP source to native binary.
 /// Reads PHP source, tokenizes, parses, resolves names, type-checks, optimizes,
@@ -83,59 +82,7 @@ pub(crate) fn compile(config: CliConfig) {
     let output_paths = output_paths(filename, target, emit);
     let mut timings = CompileTimings::new(emit_timings);
 
-    crate::progress::phase("read");
-    let phase_started = Instant::now();
-    let source = match fs::read_to_string(filename) {
-        Ok(s) => s,
-        Err(e) => {
-            crate::progress::clear();
-            eprintln!("Error reading '{}': {}", filename, e);
-            process::exit(1);
-        }
-    };
-    timings.record_since("read", phase_started);
-
-    crate::progress::phase("tokenize");
-    let phase_started = Instant::now();
-    let tokens = match lexer::tokenize_with_mode(&source, source_mode) {
-        Ok(tokens) => tokens,
-        Err(e) => {
-            crate::progress::clear();
-            errors::report(&e.with_file(filename.to_string()));
-            process::exit(1);
-        }
-    };
-    timings.record_since("tokenize", phase_started);
-
-    crate::progress::phase("parse");
-    let phase_started = Instant::now();
-    let parsed = match parser::parse_with_mode(&tokens, source_mode) {
-        Ok(ast) => ast,
-        Err(e) => {
-            crate::progress::clear();
-            errors::report(&e.with_file(filename.to_string()));
-            process::exit(1);
-        }
-    };
-    timings.record_since("parse", phase_started);
-
-    crate::progress::phase("magic-constants");
-    let phase_started = Instant::now();
-    let main_file_path = Path::new(filename).to_path_buf();
-    let parsed = match crate::source::finalize_physical_program(
-        parsed,
-        &main_file_path,
-        source_mode,
-        &defines,
-    ) {
-        Ok(parsed) => parsed,
-        Err(e) => {
-            crate::progress::clear();
-            errors::report(&e);
-            process::exit(1);
-        }
-    };
-    timings.record_since("magic-constants", phase_started);
+    let parsed = frontend::read_and_parse(filename, source_mode, &defines, &mut timings);
 
     crate::progress::phase("autoload-build");
     let phase_started = Instant::now();
@@ -484,41 +431,17 @@ pub(crate) fn compile(config: CliConfig) {
     timings.record_since("dce", phase_started);
 
     if emit_ir {
-        crate::progress::phase("ir-lower");
-        let phase_started = Instant::now();
-        let mut module = match ir_lower::lower_program_with_source_path_and_web(
+        eir_output::emit(
             &ast,
             &check_result,
             target,
-            Path::new(filename),
+            filename,
             web,
-        ) {
-            Ok(module) => module,
-            Err(err) => {
-                crate::progress::clear();
-                eprintln!("EIR lowering error: {}", err);
-                process::exit(1);
-            }
-        };
-        timings.record_since("ir-lower", phase_started);
-
-        crate::progress::phase("ir-opt");
-        let phase_started = Instant::now();
-        if ir_opt {
-            ir_passes::optimize_module(&mut module);
-        }
-        timings.record_since("ir-opt", phase_started);
-
-        crate::progress::phase("ir-print");
-        let phase_started = Instant::now();
-        let text = ir::print_module(&module);
-        timings.record_since("ir-print", phase_started);
-        crate::progress::clear();
-        timings.report();
-        print!("{}", text);
+            ir_opt,
+            &mut timings,
+        );
         return;
     }
-
     crate::progress::phase("ir-lower");
     let phase_started = Instant::now();
     let mut ir_module = match ir_lower::lower_program_with_source_path_and_web(
@@ -544,278 +467,26 @@ pub(crate) fn compile(config: CliConfig) {
     }
     timings.record_since("ir-opt", phase_started);
 
-    // Dynamic source is opaque to AOT feature detection. `--with-regex`
-    // explicitly enables the ordinary regex runtime/native requirement and
-    // lets eval setup register that managed provider with Magician.
-    if with_crates.contains("regex") {
-        ir_module.required_runtime_features.regex = true;
-    }
-    let mut runtime_features = ir_module.required_runtime_features;
-    // `--web` selects the output-capture variant of `__rt_stdout_write`. This is the
-    // sole driver of the web runtime feature: it is CLI-driven, not derived from the
-    // program, so the runtime cache (keyed on the generated assembly hash) keeps the
-    // web and non-web runtime objects distinct automatically.
-    runtime_features.web = web;
-
-    let runtime_link_requirements =
-        codegen::link_requirements_for_runtime_features(runtime_features);
-
-    // Bridge-backed `--with-<name>` values force-link their staticlib
-    // (whole-archived via `forced_bridge_libs`) regardless of feature
-    // auto-detection. Runtime-only capabilities such as regex have already
-    // updated `runtime_features` and intentionally do not map to a bridge.
-    let mut forced_bridge_libs: Vec<String> = Vec::new();
-    let mut sorted_with_crates: Vec<&String> = with_crates.iter().collect();
-    sorted_with_crates.sort();
-    for flag in sorted_with_crates {
-        if let Some(lib) = linker::bridge_lib_for_flag(flag) {
-            forced_bridge_libs.push(lib.to_string());
-        }
-    }
-
-    // Collect the named libraries that the typed link planner will consider.
-    // This preserves codegen-time bridge feature reporting without flattening
-    // those inputs back into the legacy user-library list.
-    let mut planned_link_libraries = Vec::new();
-    for library in extra_link_libs
-        .iter()
-        .chain(check_result.required_libraries.iter())
-        .chain(forced_bridge_libs.iter())
-    {
-        if !planned_link_libraries.contains(library) {
-            planned_link_libraries.push(library.clone());
-        }
-    }
-    if web && !planned_link_libraries.iter().any(|library| library == "elephc_web") {
-        planned_link_libraries.push("elephc_web".to_string());
-    }
-    for requirement in &runtime_link_requirements {
-        if let LinkRequirement::Bridge(library) = requirement {
-            if !planned_link_libraries
-                .iter()
-                .any(|existing| existing.as_str() == *library)
-            {
-                planned_link_libraries.push((*library).to_string());
-            }
-        }
-    }
-
-    let requires_elephc_tls = planned_link_libraries
-        .iter()
-        .any(|library| library == "elephc_tls");
-
-    // Report the bridges actually linked into THIS compilation to
-    // `extension_loaded()` / `get_loaded_extensions()`. Each planned bridge is
-    // mapped through the single-source bridge table; bridges with no distinct
-    // PHP extension (tz -> date, eval) are skipped. Seeded into a codegen
-    // thread-local because extension folding happens during instruction lowering.
-    let mut linked_extensions: Vec<String> = Vec::new();
-    for lib in &planned_link_libraries {
-        if let Some(ext) = linker::php_extension_for_lib(lib) {
-            if !linked_extensions.iter().any(|existing| existing == ext) {
-                linked_extensions.push(ext.to_string());
-            }
-        }
-    }
-    codegen::set_linked_extensions(linked_extensions);
-
-    crate::progress::phase("runtime-cache");
-    let phase_started = Instant::now();
-    let runtime_pic = matches!(emit, Emit::Cdylib);
-    let runtime_object = match runtime_cache::prepare_runtime_object(heap_size, target, runtime_features, runtime_pic) {
-        Ok(runtime_object) => runtime_object,
-        Err(err) => {
-            crate::progress::clear();
-            eprintln!("Runtime cache error: {}", err);
-            process::exit(1);
-        }
-    };
-    timings.record_since("runtime-cache", phase_started);
-    timings.note(format!("Runtime cache: {}", runtime_object.status.as_str()));
-
-    crate::progress::phase("codegen");
-    let phase_started = Instant::now();
-    let user_asm = match codegen::generate_user_asm_from_ir_with_options(
-        &ir_module,
+    backend::emit_and_link(backend::BackendInputs {
+        filename,
+        with_crates: &with_crates,
+        ir_module,
+        web,
+        extra_link_libs: &extra_link_libs,
+        extra_link_paths: &extra_link_paths,
+        extra_frameworks: &extra_frameworks,
+        required_libraries: &check_result.required_libraries,
+        target,
+        emit,
+        heap_size,
         gc_stats,
         heap_debug,
-        requires_elephc_tls,
-        emit,
-        &exported_functions,
+        exported_functions: &exported_functions,
         regalloc_linear,
-        web,
-    ) {
-        Ok(asm) => asm,
-        Err(err) => {
-            crate::progress::clear();
-            eprintln!("EIR backend error: {}", err);
-            process::exit(1);
-        }
-    };
-    let user_asm = if emit_debug_info {
-        debug_info::inject_line_directives(&user_asm, filename, target.platform)
-    } else {
-        user_asm
-    };
-    timings.record_since("codegen", phase_started);
-
-    crate::progress::phase("write-asm");
-    let phase_started = Instant::now();
-    if let Err(e) = fs::write(&output_paths.asm, &user_asm) {
-        crate::progress::clear();
-        eprintln!("Error writing '{}': {}", output_paths.asm.display(), e);
-        process::exit(1);
-    }
-    timings.record_since("write-asm", phase_started);
-
-    if emit_source_map {
-        crate::progress::phase("source-map");
-        let phase_started = Instant::now();
-        if let Err(err) =
-            source_map::write_source_map(
-                &user_asm,
-                Path::new(filename),
-                &output_paths.asm,
-                &output_paths.source_map,
-            )
-        {
-            crate::progress::clear();
-            eprintln!("Source map error: {}", err);
-            process::exit(1);
-        }
-        timings.record_since("source-map", phase_started);
-    }
-
-    if emit_asm {
-        crate::progress::clear();
-        timings.report();
-        crate::progress::finish_ok(
-            &format!(
-                "Emitted assembly '{}' -> '{}'",
-                filename,
-                output_paths.asm.display()
-            ),
-            timings.elapsed(),
-        );
-        return;
-    }
-
-    let native_requirements: Vec<NativeRequirement> = runtime_link_requirements
-        .iter()
-        .filter_map(|requirement| match requirement {
-            LinkRequirement::NativePackage(package) => {
-                Some(NativeRequirement::package(*package))
-            }
-            LinkRequirement::Bridge(_) | LinkRequirement::SystemLibrary(_) => None,
-        })
-        .collect();
-    let resolved_native = match crate::native_deps::resolve_for_compilation(
-        Path::new(filename),
-        target,
-        &native_requirements,
-    ) {
-        Ok(resolved) => resolved,
-        Err(error) => {
-            eprintln!("{error}");
-            process::exit(1);
-        }
-    };
-    let link_plan = crate::link_planning::build(crate::link_planning::LinkPlanningInputs {
-        user_libraries: &extra_link_libs,
-        user_search_paths: &extra_link_paths,
-        user_frameworks: &extra_frameworks,
-        checker_libraries: &check_result.required_libraries,
-        runtime_requirements: &runtime_link_requirements,
-        managed_packages: &resolved_native,
-        forced_bridges: &forced_bridge_libs,
-        web,
+        emit_debug_info,
+        output_paths: &output_paths,
+        emit_source_map,
+        emit_asm,
+        timings: &mut timings,
     });
-
-    crate::progress::phase("assemble");
-    let phase_started = Instant::now();
-    linker::assemble(target, &output_paths.asm, &output_paths.obj);
-    timings.record_since("assemble", phase_started);
-
-    for (lib_name, flag_name) in linker::bridges_in(&planned_link_libraries) {
-        let detail = if forced_bridge_libs.iter().any(|l| l == lib_name) {
-            format!("{} (--with-{})", lib_name, flag_name)
-        } else {
-            format!("{} (auto-detected)", lib_name)
-        };
-        crate::progress::event("Linking", &detail);
-    }
-
-    crate::progress::phase("link");
-    let phase_started = Instant::now();
-    if let Err(error) = linker::link_with_plan(
-        target,
-        emit,
-        &output_paths.bin,
-        &output_paths.obj,
-        &runtime_object.path,
-        &link_plan,
-        &forced_bridge_libs,
-    ) {
-        eprintln!("Linker error: {error}");
-        process::exit(1);
-    }
-    timings.record_since("link", phase_started);
-
-    // With --debug-info the DWARF line tables must be preserved past object
-    // cleanup: on macOS `dsymutil` bakes them into a .dSYM while the object
-    // still exists; if that fails the object is kept so debuggers can follow
-    // the binary's debug map to it.
-    let keep_obj_for_debug =
-        emit_debug_info && !linker::bake_debug_info(target, &output_paths.bin);
-    if !keep_obj_for_debug {
-        let _ = fs::remove_file(&output_paths.obj);
-    }
-
-    crate::progress::clear();
-    timings.report();
-    if let Some(warning) = dynamic_eval_capability_warning(runtime_features) {
-        eprintln!("{warning}");
-    }
-    crate::progress::finish_ok(
-        &format!("Compiled '{}' -> '{}'", filename, output_paths.bin.display()),
-        timings.elapsed(),
-    );
-}
-
-/// Returns the post-link reminder for dynamic eval without optional regex support.
-fn dynamic_eval_capability_warning(
-    runtime_features: codegen::RuntimeFeatures,
-) -> Option<&'static str> {
-    (runtime_features.eval_bridge && !runtime_features.regex).then_some(concat!(
-        "warning: dynamic eval was compiled without optional regex support\n",
-        "evaluated code that uses preg_* or mb_ereg_match() will fail at runtime; enable it with:\n",
-        "  elephc native add pcre2\n",
-        "  elephc --with-regex <source-file>",
-    ))
-}
-
-/// Computes output paths for .s (assembly), .o (object), binary, and .map (source map) files
-/// derived from the input filename.
-///
-/// Executable mode produces `<stem>` (no extension). Cdylib mode produces
-/// `lib<stem>.so` (Linux) or `lib<stem>.dylib` (macOS), matching the conventional
-/// shared-library naming that `dlopen(3)` and linker `-l` flags expect.
-fn output_paths(filename: &str, target: Target, emit: Emit) -> OutputPaths {
-    let path = Path::new(filename);
-    let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("output");
-    let parent = path.parent().unwrap_or(Path::new("."));
-    let bin_name = match emit {
-        Emit::Executable => stem.to_string(),
-        Emit::Cdylib => match target.platform {
-            Platform::MacOS => format!("lib{}.dylib", stem),
-            Platform::Linux => format!("lib{}.so", stem),
-            Platform::Windows => panic!("Windows target is not yet supported (see issue #379)"),
-        },
-    };
-    OutputPaths {
-        asm: parent.join(format!("{}.s", stem)),
-        obj: parent.join(format!("{}.o", stem)),
-        bin: parent.join(bin_name),
-        source_map: parent.join(format!("{}.map", stem)),
-    }
 }
